Refactored Version class into version.py
